+++ src/components/layout/Headers/EditHeader/index.tsx
@@ -34,7 +34,7 @@ import {
   getTag,
 } from '@/graphql/queries';
 import { AuthContext } from '@/pages/_app';
-import { tag } from '@/apiHelper';
+import { tag, image } from '@/apiHelper';
 
 import styles from './style.module.scss';
 
@@ -86,73 +86,9 @@ export default function EditHeader({ editorState, postId }) {
 
     await tag.deleteAndUnLinkLegacyTag({ tags, postTagsInDB, postId });
 
-    // Get DB images
-    const listImagesRes = await API.graphql({
-      query: listImages,
-      variables: {
-        filter: { isPublished: { eq: false } },
-      },
-    });
-    const draftImagesDB = listImagesRes.data.listImages.items;
-    const imagesToDelete = draftImagesDB.filter(
-      (imageDB) => !imageKeys.find((imageKey) => imageKey === imageDB.imageKey)
-    );
+    await image.trimImageS3AndDB({ postId, images });
 
-    //  Delete Image on S3 and DB if not exist now
-    for (const imageToDelete of imagesToDelete) {
-      const delRes = await Storage.remove(imageToDelete.imageKey);
-      console.log(`delRes`, delRes);
-
-      // Delete connection
-      const postImageRes = await API.graphql({
-        query: postImageByPostIdAndImageId,
-        variables: { postId, imageId: { eq: imageToDelete.id } },
-      });
-      const postImageId =
-        postImageRes?.data?.postImageByPostIdAndImageId?.items[0].id ?? null;
-      if (postImageId) {
-        await API.graphql({
-          query: deletePostImage,
-          variables: { input: { id: postImageId } },
-        });
-      }
-
-      const deletedRes = await API.graphql({
-        query: deleteImage,
-        variables: { input: { id: imageToDelete.id } },
-      });
-      console.log(`deletedDBRes`, deletedRes);
-    }
-
-    // Get PostImages on this Post
-
-    const postImagesDBRes = await API.graphql({
-      query: listPostImages,
-      variables: { filter: { postId: { eq: postId } } },
-    });
-    const postImageDB = postImagesDBRes?.data?.listPostTags?.items ?? [];
-
-    // Create Image Mapping if not mapped already
-    for (const image of images) {
-      const isMapedDBAlready = !!postImageDB.find(
-        (imageDB) => imageDB.id === image.data.imageDbId
-      );
-      if (!isMapedDBAlready) {
-        const createPostImageRes = await API.graphql({
-          query: createPostImage,
-          variables: {
-            input: {
-              postId,
-              userId,
-              imageId: image.data.imageDbId,
-              baseType: 'PostImage',
-            },
-          },
-        });
-
-        console.log(`createPostImageRes`, createPostImageRes);
-      }
-    }
+    await image.mapPostAndIamges({ postId, userId, images });
 
     return editedPostId;
   };
