+++ src/components/layout/Headers/PostHeader/index.tsx
@@ -20,7 +20,7 @@ import {
   deleteImage,
 } from '@/graphql/mutations';
 import { AuthContext } from '@/pages/_app';
-import { tag } from '@/apiHelper';
+import { tag, image } from '@/apiHelper';
 
 import styles from './style.module.scss';
 
@@ -67,63 +67,20 @@ export default function PostHeader({ editorState, owner, post }) {
           onClick={async () => {
             setShowModal(false);
             const postTagsInDB = await tag.getTagsByPostId({ postId });
-            console.log(`postTagsInDB`, postTagsInDB);
+
             await tag.deleteAndUnLinkLegacyTag({
               tags: [],
               postTagsInDB,
               postId,
             });
 
+            await image.deleteImageS3AndDB({ post });
+
             // Delete Post
             await API.graphql({
               query: deletePost,
               variables: { input: { id: postId } },
             });
-
-            // // Delete Post Tags
-            // const postTags = post.postTags.items;
-            // for (const postTag of postTags) {
-            //   const deletePostTagRes = await API.graphql({
-            //     query: deletePostTag,
-            //     variables: { input: { id: postTag.id } },
-            //   });
-            //   console.log(`deletePostTagRes`, deletePostTagRes);
-            //   const deletedLinkTag = deletePostTagRes.data.deletePostTag.tag;
-
-            //   //  If there is no more linked post delete tag
-            //   console.log(`deletedLinkTag`, deletedLinkTag);
-            //   if (deletedLinkTag.postTags.nextToken === null) {
-            //     const deletedTagRes = await API.graphql({
-            //       query: deleteTag,
-            //       variables: { input: { id: deletedLinkTag.id } },
-            //     });
-            //     console.log(`deletedTagRes`, deletedTagRes);
-            //   }
-            // }
-
-            // Delete Post Images
-            const postImages = post.postImages.items;
-            for (const postImage of postImages) {
-              const deletePostImageRes = await API.graphql({
-                query: deletePostImage,
-                variables: { input: { id: postImage.id } },
-              });
-              console.log(`deletePostImageRes`, deletePostImageRes);
-              const deletedLinkImage =
-                deletePostImageRes.data.deletePostImage.image;
-              // If There is no more linked post delete image
-              if (deletedLinkImage.postImage.nextToken === null) {
-                const storageDeleteRes = await Storage.remove(
-                  deletedLinkImage.imageKey
-                );
-                console.log(`storageDeleteRes`, storageDeleteRes);
-                const dbImageDeleteRes = await API.graphql({
-                  query: deleteImage,
-                  variables: { input: { id: deletedLinkImage.id } },
-                });
-                console.log(`dbImageDeleteRes`, dbImageDeleteRes);
-              }
-            }
             router.push('/');
           }}
         >
